Skip the preflight check
gh-1

Whatever problem the preflight check that I'm disabling here is trying
ot solve, I bet it doesn't apply because frontend/ is the top-level
directory when the frontend artifact gets built.

I am skipping the preflight check to fix this failure in my build:

There might be a problem with the project dependency tree.
It is likely not a bug in Create React App, but something you need to
fix locally.

The react-scripts package provided by Create React App requires a
dependency:

  "babel-jest": "^26.6.0"

Don't try to install it manually: your package manager does it
automatically.
However, a different version of babel-jest was detected higher up in the
tree:

  /home/runner/work/room-app/room-app/node_modules/babel-jest (version:
27.0.2)

Manually installing incompatible versions is known to cause
hard-to-debug issues.

If you would prefer to ignore this check, add SKIP_PREFLIGHT_CHECK=true
to an .env file in your project.
That will permanently disable this message but you might encounter other
issues.